+++ crates/trading/src/python/mod.rs
@@ -44,6 +44,8 @@ pub fn trading(_: Python<'_>, m: &Bound<'_, PyModule>) -> PyResult<()> {
     m.add_class::<crate::strategy::ImportableStrategyConfig>()?;
     m.add_class::<crate::algorithm::ImportableExecAlgorithmConfig>()?;
     #[cfg(feature = "examples")]
+    m.add_class::<crate::examples::strategies::CompositeMarketMakerConfig>()?;
+    #[cfg(feature = "examples")]
     m.add_class::<crate::examples::strategies::EmaCrossConfig>()?;
     #[cfg(feature = "examples")]
     m.add_class::<crate::examples::strategies::GridMarketMakerConfig>()?;